Change analysis version interface for API
Before when creating an analysis information with a request you would
need to specify in the request body:

{..., "version": <version_pk>, ...}

That's a bit annoying because you have to look up the PK for the
analysis version if you don't already have it written down.

And if the analysis version didn't exist, then you had to create it
manually using the admin interface.

That's a lot of work. Sophia suggested just being able to pass in the
analysis version as a string, and there's a lot of sense to that. That's
what this commit implements.

Changes to the API for GET requests are that instead of the nested

"version": {"version": "Single Cell Pipeline v0_0_6"}

we now just have

"version": "Single Cell Pipeline v0_0_6"

And when you make a POST request you give it a string instead of the PK.
If the analysis version matching the string doesn't exist, then it
creates an analysis version with that version string; otherwise it just
uses the existing version matching the string passed in.

Downsides to this approach is that it doesn't protect against multiple
synonyms of the same version; for example, it doesn't protect against
users passing in

0.0.6 -or- scp_0.0.6 -or- Single Cell Pipeline v0_0_6

and will treat each of these versions as distinct. So there's a bit of a
tradeoff (strictness for convenience), but the convenience is probably
worth it here.